Deduce PS2 region from Redumper serial
