+++ MPF.Frontend/Drive.cs
@@ -10,6 +10,7 @@ using Microsoft.Management.Infrastructure.Generic;
 #endif
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using MPF.Processors;
 using SabreTools.IO;
 using SabreTools.RedumpLib.Data;
 
@@ -328,7 +329,7 @@ namespace MPF.Frontend
                 .Replace(".", string.Empty);
 
             // Get the region, if possible
-            region = GetPlayStationRegion(exeName);
+            region = ProcessingTool.GetPlayStationRegion(exeName);
 
             // Now that we have the EXE name, try to get the fileinfo for it
             string exePath = Path.Combine(Name, exeName);
@@ -342,91 +343,6 @@ namespace MPF.Frontend
             date = dt.ToString("yyyy-MM-dd");
 
             return true;
-        }
-
-        /// <summary>
-        /// Determine the region based on the PlayStation serial code
-        /// </summary>
-        /// <param name="serial">PlayStation serial code</param>
-        /// <returns>Region mapped from name, if possible</returns>
-        public static Region? GetPlayStationRegion(string? serial)
-        {
-            // If we have a fully invalid serial
-            if (string.IsNullOrEmpty(serial))
-                return null;
-
-            // Standardized "S" serials
-            if (serial!.StartsWith("S"))
-            {
-                // string publisher = serial[0] + serial[1];
-                // char secondRegion = serial[3];
-                switch (serial[2])
-                {
-                    case 'A': return Region.Asia;
-                    case 'C': return Region.China;
-                    case 'E': return Region.Europe;
-                    case 'K': return Region.SouthKorea;
-                    case 'U': return Region.UnitedStatesOfAmerica;
-                    case 'P':
-                        // Region of S_P_ serials may be Japan, Asia, or SouthKorea
-                        return serial[3] switch
-                        {
-                            // Check first two digits of S_PS serial
-                            'S' => (Region?)(serial.Substring(5, 2) switch
-                            {
-                                "46" => Region.SouthKorea,
-                                "51" => Region.Asia,
-                                "56" => Region.SouthKorea,
-                                "55" => Region.Asia,
-                                _ => Region.Japan,
-                            }),
-
-                            // Check first three digits of S_PM serial
-                            'M' => (Region?)(serial.Substring(5, 3) switch
-                            {
-                                "645" => Region.SouthKorea,
-                                "675" => Region.SouthKorea,
-                                "885" => Region.SouthKorea,
-                                _ => Region.Japan, // Remaining S_PM serials may be Japan or Asia
-                            }),
-                            _ => (Region?)Region.Japan,
-                        };
-                }
-            }
-
-            // Japan-only special serial
-            else if (serial.StartsWith("PAPX"))
-                return Region.Japan;
-
-            // Region appears entirely random
-            else if (serial.StartsWith("PABX"))
-                return null;
-
-            // Region appears entirely random
-            else if (serial.StartsWith("PBPX"))
-                return null;
-
-            // Japan-only special serial
-            else if (serial.StartsWith("PCBX"))
-                return Region.Japan;
-
-            // Japan-only special serial
-            else if (serial.StartsWith("PCXC"))
-                return Region.Japan;
-
-            // Single disc known, Japan
-            else if (serial.StartsWith("PDBX"))
-                return Region.Japan;
-
-            // Single disc known, Europe
-            else if (serial.StartsWith("PEBX"))
-                return Region.Europe;
-
-            // Single disc known, USA
-            else if (serial.StartsWith("PUBX"))
-                return Region.UnitedStatesOfAmerica;
-
-            return null;
         }
 
         /// <summary>
